+++ src/main/java/com/devloop/lecture/response/GetLectureListResponse.java
@@ -3,6 +3,8 @@ package com.devloop.lecture.response;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @NoArgsConstructor
 public class GetLectureListResponse {
@@ -10,18 +12,16 @@ public class GetLectureListResponse {
     private String title;
     private String category;
     private String level;
-    private Integer price;
+    private BigDecimal price;
 
-    //튜터 이름
-    //후기 평균
-    private GetLectureListResponse(Long id, String title, String category, String level, Integer price){
+    private GetLectureListResponse(Long id, String title, String category, String level, BigDecimal price){
         this.id=id;
         this.title=title;
         this.category=category;
         this.level=level;
         this.price=price;
     }
-    public static GetLectureListResponse of(Long id, String title, String category, String level, Integer price){
+    public static GetLectureListResponse of(Long id, String title, String category, String level, BigDecimal price){
         return new GetLectureListResponse(
                 id,
                 title,
